Enable the n dimensional parallelization (#170)
Summary:
Pull Request resolved: https://github.com/pytorch/FBGEMM/pull/170

Enabling the parallelization on the n dimension in FBGEMM. This is useful when we have the multi-thread Predictor and the target matrix shapes have large n dimension sizes. With this Diff, we can also parallelize both m and n dimensions (2D parallelization) simultaneously.

When parallelizing the n dimension, note that we have the additional overhead of allocating the duplicated packA buffer for each thread.

Reviewed By: ilia-cher

Differential Revision: D17696021

fbshipit-source-id: cd20772d800481ec42fec572499dd20cceb7170b