Fix the _aligned_free() issue on Windows
Summary:
When we run `PackedRequantizeTest` on Windows, it shows the following error:
"The Block at XXX was allocated by aligned routines, use _aligned_free()".

{F227468367}

The root cause should be the inappropriate allocation API (not Windows compatible) in the Packing routines in FBGEMM. This Diff provides a quick fix in the unit tests (Previously our tests also had bugs as ` row_offset_buf` is declared but not used.

Differential Revision: D19659649

fbshipit-source-id: 455ebd8641315fbedb9b0ea5d9148be8ab92c4be