More fixes to enable USE_FBGEMM for PyTorch on Windows (#291)
Summary:
Pull Request resolved: https://github.com/pytorch/FBGEMM/pull/291

Pull Request resolved: https://github.com/pytorch/pytorch/pull/33000

As suggested by  peterjc123 in
https://github.com/pytorch/pytorch/pull/33000#issuecomment-582893875, we would like to fix the configuration of the PyTorch CMake scripts by modifying the `CMakeList.txt` at FBGEMM.

In addition, we add more `FBGEMM_API` instances to address the issues on using dllimport and dllexport in C++ classes on Windows.

We split FBGEMM changes as a separate Diff so that PyTorch master branch will not be broken due to the asynchronization between FBGEMM and PyTorch.
ghstack-source-id: 97914400

Differential Revision: D19786708

fbshipit-source-id: 38329a4371d5f6be0ecf9c291fa1ee161d92fc9c